+++ ext/custom-transaction-processes/default-purchase/process.edn
@@ -52,10 +52,6 @@
  :notifications
  [{:name :notification/order-receipt,
    :on :transition/confirm-payment,
-   :at
-   {:fn/plus
-    [{:fn/timepoint [:time/first-entered-state :state/purchased]}
-     {:fn/period ["PT15M"]}]},
    :to :actor.role/customer,
    :template :purchase-order-receipt}
   {:name :notification/purchase-new-order,
